VMMCP: FIRST COMMIT
Couple things here.

We establish the principle that vmm is always in there. It's not optional.

This ensures that everything always builds.

We don't conditionally compile in anything. If we want variant behavior,
we dynamically key on cpuinfo. It's been learned the hard way that this is the
way to go.

Signed-off-by: Ron Minnich <rminnich@gmail.com>