Add placeholder handling to devgen.
The current devgen assumes that the dirtab is static, which until now it
has been. However, the gen functions are allowed to return 0 if the
entry exists for that 's' but shouldn't be active.

Overload dirtab->qid.vers to indicate if the dirtab entry is inactive
and return 0 if vers == -1.