Hacky virtioconsole
This actually works if we don't try to read the console.

The interrupts for the console are eaten by the vmexit. Need to learn how to
kick off the interrupt processing when the vm exits with an interrupt, OR
set up cores with no interrupts going to them on which to run VMs.

Next steps are to remove remaining hacks and then see how much further
we can get with virtio, then it's on to linux.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>