Virtio programmed io console.
Starts a thread which pushes characters to the guest via a programmed io
mailbox. This is intended to be used to bootstrap a guest console
only until they are ready to use a virtio console. It should not be used
past that point.

Weirdly, getchar() in the thread is not echoing characters. This is actually
good, but weird. I figure it's another case of us not have a real console yet.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>