Pthread cleanup stubs
These cleanup stubs are similar to our waserror/poperror.  They interact
with pthread_cancel and exit.  If a thread exits or is cancelled, it's
cleanup functions are called.  Additionally, cleanup_pop takes a bool,
and when it's TRUE, it'll run the handler even when returning normally.

If someone wants to implement these and pthread_cancel, be my guest.