Pthread conds and broadcasts use SLISTS
This fixes PTHREAD_COND_INITIALIZER.  The issue was that we can't use
TAILQs, because then the static initializer would need a parameter.

Now, there's an SLIST and a TAILQ.  The 2LS lists are TAILQs.  The
cond/barrier/mutex (eventually) are SLISTs.  A thread should only be on
one at a time.

This commit also cleans up things a bit, since broadcast and barrier had
similar, repeated code.