Removes BSD shims that are (better) in glibc
We have a problem with our BSD shims.  Many of them implement functions
that are already in glibc, and we can have link errors.  The functions
removed in this commit appear to do the same thing as their glibc
counterparts, so removing them is not a problem.

Putenv is a little different.  It's not the same functionality as
glibc's, but we're using glibc-style environments and whatnot already,
so we might as well use that.

The bigger issue is the other functions:

	user/bsd/gethostbyaddr.c
	user/bsd/gethostbyname.c
	user/bsd/getprotobyname.c
	user/bsd/getservbyname.c

These appear to have a different implementation.  Gethostby all use cs
under the hood, which is different than glibc.  The long term plan is to
probably push those functions into glibc.

The other set of issues is that functions like listen and accept have
different meanings between glibc and plan 9.  Good times.