x86: changes backtrace to not subtract 1
That -1 was there for a long time, due to weird issues with our earliest
bts in some corner cases in the kernel.  The tradeoff is that the
addresses given aren't accurate (off by 1!), and that make manual symbol
lookup a little annoying.  For instance, using a user-backtrace, you
can't grep an objdump for a particular PC, since the -1 means the
address probably doesn't exist.